+++ .teamcity/Testbench/RegressionTestODESolve/RegressionTestODESolve.kt
@@ -1,4 +1,4 @@
-package Testbench
+package Testbench.RegressionTestODESolve
 
 import Ribasim_Windows.Windows_BuildRibasim
 import Ribasim_Linux.Linux_BuildRibasim
